Rename class LogLine to module LogLineParser
